// src/models/evaluate-request.ts
/* tslint:disable */
/* eslint-disable */
/**
 * flexCharge-service
 * No description provided (generated by Swagger Codegen https://github.com/swagger-api/swagger-codegen)
 *
 * OpenAPI spec version: v1
 *
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */
import { AdditionalField } from './additional-field';
import { BillingInformation } from './billing-information';
import { Merchant } from './merchant';
import { OrderItem } from './order-item';
import { Payer } from './payer';
import { PaymentMethod } from './payment-method';
import { ShippingInformation } from './shipping-information';
import { Subscription } from './subscription';
import { Transaction } from './transaction';
/**
 *
 * @export
 * @interface EvaluateRequest
 */
export interface EvaluateRequest {
    /**
     *
     * @type {boolean}
     * @memberof EvaluateRequest
     */
    isDeclined: boolean;
    /**
     *
     * @type {string}
     * @memberof EvaluateRequest
     */
    mid?: string | null;
    /**
     *
     * @type {string}
     * @memberof EvaluateRequest
     */
    orderId: string;
    /**
     *
     * @type {string}
     * @memberof EvaluateRequest
     */
    orderSource?: string | null;
    /**
     *
     * @type {Transaction}
     * @memberof EvaluateRequest
     */
    transaction: Transaction;
    /**
     *
     * @type {Payer}
     * @memberof EvaluateRequest
     */
    payer: Payer;
    /**
     *
     * @type {Array<OrderItem>}
     * @memberof EvaluateRequest
     */
    orderItems?: Array<OrderItem> | null;
    /**
     *
     * @type {ShippingInformation}
     * @memberof EvaluateRequest
     */
    shippingInformation?: ShippingInformation;
    /**
     *
     * @type {boolean}
     * @memberof EvaluateRequest
     */
    isMIT?: boolean | null;
    /**
     *
     * @type {boolean}
     * @memberof EvaluateRequest
     */
    isRecurring?: boolean | null;
    /**
     *
     * @type {Subscription}
     * @memberof EvaluateRequest
     */
    subscription?: Subscription | null;
    /**
     *
     * @type {Array<AdditionalField>}
     * @memberof EvaluateRequest
     */
    additionalFields?: Array<AdditionalField> | null;
    /**
     *
     * @type {string}
     * @memberof EvaluateRequest
     */
    correlationId?: string;
    /**
     *
     * @type {string}
     * @memberof EvaluateRequest
     */
    idempotencyKey: string;
    /**
     *
     * @type {string}
     * @memberof EvaluateRequest
     */
    senseKey?: string | null;
    /**
     *
     * @type {Merchant}
     * @memberof EvaluateRequest
     */
    merchant?: Merchant;
    /**
     *
     * @type {PaymentMethod}
     * @memberof EvaluateRequest
     */
    paymentMethod?: PaymentMethod;
    /**
     *
     * @type {BillingInformation}
     * @memberof EvaluateRequest
     */
    billingInformation: BillingInformation;
    /**
     *
     * @type {Date}
     * @memberof EvaluateRequest
     */
    expiryDateUtc?: Date | string | null;
}